+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Features:
 
 - Add back support for building serverless functions.
+- Use the first enum case of the enum hint as the default value in generated test objects.
 
 ## v0.12.0 (2025-08-28)